% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{pwr1}
\alias{pwr1}
\title{Power Function}
\usage{
pwr1(x = 2, y = 3)
}
\arguments{
\item{x}{Base value}

\item{y}{The exponent}
}
\value{
numerical value or a message
}
\description{
Power Function
}
\details{
It allowed you to raise a numerical value to twice of the other and if there is a non-numeric value it will tell you
}
\examples{
pwr1()
pwr1(x = 2, y =5)
pwr1(2, "a")
}
\author{
Hamid Semiyari
}
