% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{pwr2}
\alias{pwr2}
\title{Create column Power of a data frame}
\usage{
pwr2(x = c(2, 2), y = c(3, 5))
}
\arguments{
\item{x}{is a vector}

\item{y}{is a vector}
}
\value{
a datafram
}
\description{
Two numeric vectors are creating a new vector and it will return a tibble
}
\examples{
pwr2()
pwr2(x=c(-2,3, 10), y=c(3,5, -1))

}
